Heuristic uses double instead of int

+++ src/Heuristic/AvoidGameOver.java
@@ -5,7 +5,7 @@ import tetrisGame.TetrisMap;
 public class AvoidGameOver extends Heuristic {
 
 	@Override
-	public int calculate(TetrisMap gm) {
+	public double calculate(TetrisMap gm) {
 		return gm.isGameOver()?super.weight:0;
 	}
