+++ src/Heuristic/Diagonal.java
@@ -10,7 +10,7 @@ import tetrisGame.TetrisMap;
 public class Diagonal extends Heuristic{
 
 	@Override
-	public int calculate(TetrisMap gm) {
+	public double calculate(TetrisMap gm) {
 		int leftHeight = gm.map.get(0).size();
 		int count= 0;
 		for(int i=1; i< gm.col; ++i){
